modify related references
